pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agents-inc"
version = "4.0.0"
description = "Publication-ready multi-agent group fabric and router orchestration toolkit for Codex sessions"
readme = "README.md"
requires-python = ">=3.8"
authors = [
  {name = "sacRedeeRhoRn"}
]
license = {text = "MIT"}
keywords = ["codex", "multi-agent", "orchestration", "skills", "research"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "black>=24.0",
  "ruff>=0.6.0"
]

[project.urls]
Homepage = "https://github.com/sacRedeeRhoRn/agents-inc"
Repository = "https://github.com/sacRedeeRhoRn/agents-inc"
Issues = "https://github.com/sacRedeeRhoRn/agents-inc/issues"

[project.scripts]
agents-inc = "agents_inc.cli.main:main"
agents-inc-skills = "agents_inc.cli.skills:main"
agents-inc-groups = "agents_inc.cli.groups:main"
agents-inc-migrate-v2 = "agents_inc.cli.migrate_v2:main"
agents-inc-orchestrate = "agents_inc.cli.orchestrate:main"
agents-inc-orchestrate-report = "agents_inc.cli.orchestrate_report:main"
agents-inc-orchestrator-reply = "agents_inc.cli.orchestrator_reply:main"
agents-inc-eval = "agents_inc.cli.eval:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agents_inc = [
  "resources/**/*",
  "docs/internal/*.md"
]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501"]
